requirements.txt
aiogram==3.5.0
motor==3.4.0
pymongo==4.7.1
python-dotenv==1.0.1